requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
fonttools==4.59.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-resources==6.5.2
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.7
matplotlib==3.9.4
nltk==3.9.1
numpy==2.0.2
openai==1.99.9
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.7.34
requests==2.32.4
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
wordcloud==1.9.4
zipp==3.23.0
